Implemented test for LocateService()
Note that running this on the Mac is problematic because MacOS
keeps shutting down portmapper/rpcbind - I guess because no
services are currently registered with it :-(.

To run this on the Mac:
  sudo launchctl load -w /System/Library/LaunchDaemons/com.apple.rpcbind.plist
  sudo launchctl start com.apple.rpcbind

+++ common.go
@@ -133,7 +133,7 @@ type RecordFragmentMarkStruct struct { //                                     Us
 }
 
 const (
-	PmapAndRpcbPort = uint32(111) // Used for Portmapper (Program Version 2) & Rpcbind (Program Versions 3 & 4)
+	PmapAndRpcbPort = uint16(111) // Used for Portmapper (Program Version 2) & Rpcbind (Program Versions 3 & 4)
 	//                               Note 1: Must actually fit in a uint16
 	//                               Note 2: Portmapper/Rpcbind typically listens on both UDP and TCP
 )
